% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcRE.R
\name{calcRE}
\alias{calcRE}
\title{Calculate relative efficiency}
\usage{
calcRE(data, rho = 0.5, method = "M_KK")
}
\arguments{
\item{data}{Input data frame}

\item{rho}{Correlation coefficient between visits}

\item{method}{Method or estimator}
}
\value{
Relative efficiency (RE) of the specified estimator, relative to the endpoint contrast (M_K)
}
\description{
Calculate relative efficiency
}
\examples{
dat2 <- padzero(sampleData)
calcRE(dat2, method="AUC")
}

