% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formData.R
\name{formData}
\alias{formData}
\title{Form a data frame based on user specified information}
\usage{
formData(visits, times, trtProfile, plaProfile)
}
\arguments{
\item{visits}{Vector of visit id variable}

\item{times}{Vector of time points corresponding to visits}

\item{trtProfile}{Vector of longitudinal change from baseline in Treatment arm (excluding baseline)}

\item{plaProfile}{Vector of longitudinal change from baseline in Placebo arm (excluding baseline)}
}
\value{
A data frame with 4 columns: visits, times, y, and trt.
}
\description{
Form a data frame based on user specified information
}
\examples{
formData(visits=1:4, times=visits, trtProfile=c(-3,rep(-4.5, length(visits)-1)),
plaProfile=c(-2,rep(-3, length(visits)-1)))
}

