% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/padzero.R
\name{padzero}
\alias{padzero}
\title{Pad zero (at baseline) to dataset}
\usage{
padzero(data)
}
\arguments{
\item{data}{Input data frame, without baseline visit}
}
\value{
Data frame with zero value padded at top
}
\description{
Pad zero (at baseline) to dataset
}
\examples{
dat <- formData(visits=1:4, times=visits, trtProfile=c(-3,rep(-4.5, length(visits)-1)),
plaProfile=c(-2,rep(-3, length(visits)-1)))
padzero(dat)
}

