% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.profiles.R
\name{plotprofiles}
\alias{plotprofiles}
\title{Plot longitudinal profiles}
\usage{
plotprofiles(dat)
}
\arguments{
\item{data}{Input data created by formData}
}
\value{
Longitudinal profile plot
}
\description{
Plot longitudinal profiles
}
\examples{
dat2 <- padzero(sampleData)
plotprofiles(dat2)
}

