% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampleData.r
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{Sample dataset.}
\format{A data frame with 4 variables}
\usage{
data(sampleData)
}
\description{
This data contains sample data.
}
\keyword{datasets}

