% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.profiles.R
\name{plotprofiles}
\alias{plotprofiles}
\title{Plot longitudinal profiles}
\usage{
plotprofiles(data)
}
\arguments{
\item{data}{Input data created by formData}
}
\value{
Longitudinal profile plot
}
\description{
Plot longitudinal profiles
}
\examples{
dat <- formData(visits=1:4, times=visits, trtProfile=c(-3,rep(-4.5, length(visits)-1)),
plaProfile=c(-2,rep(-3, length(visits)-1)))
dat2 <- padzero(dat)
plotprofiles(dat2)
}

